requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
appnope==0.1.4
ascii_colors==0.5.2
asttokens==3.0.0
attrs==25.3.0
autograd==1.7.0
beartype==0.18.5
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
configparser==7.2.0
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gensim==4.3.3
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
google-genai==1.10.0
googleapis-common-protos==1.70.0
grandalf==0.8
graphviz==0.20.3
graspologic==3.4.1
graspologic-native==1.2.5
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyppo==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
-e git+https://github.com/HKUDS/LightRAG.git@f7ee5be65838addfc6f96587832d4a9a2f97c583#egg=lightrag_hku
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
ollama==0.4.7
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
pipmaster==0.5.4
platformdirs==4.3.7
POT==0.9.5
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==78.1.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
